console: Refactor access-token module to global utility
