+++ pkg/webui/console/store/middleware/logics/user.js
@@ -16,11 +16,10 @@ import axios from 'axios'
 
 import api from '@console/api'
 
+import * as accessToken from '@ttn-lw/lib/access-token'
 import createRequestLogic from '@ttn-lw/lib/store/logics/create-request-logic'
 import { isUnauthenticatedError } from '@ttn-lw/lib/errors/utils'
 import { selectApplicationRootPath } from '@ttn-lw/lib/selectors/env'
-
-import * as accessToken from '@console/lib/access-token'
 
 import * as user from '@console/store/actions/user'
