+++ pkg/webui/lib/access-token.js
